feat: backend support for unit of measurement and decimal quantities
Database Schema:
- Add Flyway migrations V8-V11 for unit of measurement support
- Add allow_decimal, min_quantity, step_quantity columns to products table
- Change receipt_items.qty from INTEGER to DECIMAL(15,4)
- Update inventory_transactions.quantity to support decimals

Entities & DTOs:
- Update Product entity with unitOfMeasure, allowDecimal, minQuantity, stepQuantity
- Update InventoryTransaction with BigDecimal quantity support
- Update ReceiptItem with BigDecimal qty field
- Update CreateInventoryRequest, CreateProductRequest, CreateSaleRequest DTOs
- Update ProductDto and InventoryTransactionDto for decimal fields

Services:
- Update ReceiptService to handle BigDecimal quantities in sales
- Update InventoryTransactionService for decimal stock adjustments
- Update ProductService with unit of measurement validation
- Update TenantService to create schema with new unit columns
- Update UserService to include CAN_MANAGE_INVENTORY authority

Controllers:
- Update InventoryTransactionController with permission checks
- Update ProductController to handle unit of measurement data

Security:
- Add CAN_MANAGE_INVENTORY authority check for inventory operations
- Support both ROLE_ADMIN/ROLE_MANAGER and granular permissions

+++ src/main/java/com/possapp/backend/controller/ProductController.java
@@ -14,6 +14,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -99,8 +100,8 @@ public class ProductController {
             @Parameter(description = "Product code/SKU", required = true)
             @PathVariable String code,
             @Parameter(description = "Stock update request", required = true)
-            @RequestBody Map<String, Integer> request) {
-        Integer newStock = request.get("stock");
+            @RequestBody Map<String, BigDecimal> request) {
+        BigDecimal newStock = request.get("stock");
         ProductDto product = productService.updateStock(code, newStock);
         return ResponseEntity.ok(ApiResponse.success("Stock updated", product));
     }
@@ -114,8 +115,8 @@ public class ProductController {
             @Parameter(description = "Product code/SKU", required = true)
             @PathVariable String code,
             @Parameter(description = "Stock adjustment request", required = true)
-            @RequestBody Map<String, Integer> request) {
-        Integer delta = request.get("delta");
+            @RequestBody Map<String, BigDecimal> request) {
+        BigDecimal delta = request.get("delta");
         ProductDto product = productService.adjustStock(code, delta);
         return ResponseEntity.ok(ApiResponse.success("Stock adjusted", product));
     }
